Add card optimizer
We need a function which will find the best purchases to optimize the card. One that works for 2 needed
gifts and one that works for 3.

Card optimizer has two functions, one which optimizes the card for 2 gifts, one for 3. Optimizing for 2
seems efficient with nlog(n) time. 3 seems somewhat efficient at n^2 but this process will not continue
netting desirable results. Also added in testing for the card optimizer.