Creates a main method for find pair
We need a function to handle the arguments and properly format a response for this solution.

The main function parses the command line args, and strings together the processes needed to
generate the result. The prices.txt file is included as well to duplicate the results shown in
the prompt.