Create the hash digest service
We need something to generate hashes and store them as well

The hash digest service has functionality to take in messages and return hashes,
after which, it will store the pair. It also has the functionality to take in hashes
it has already produced and then return the associated message.